add ct_require macros, move ct_alloc & ct_free to alloc protocol, add Len protocol & impl for string, update test

+++ src/types/num.h
@@ -2,12 +2,12 @@
 
 #include "fatptr.h"
 
-#define $i32(x) $(Type_I32, CT_I32, 4, x)
-#define $i64(x) $(Type_I64, CT_I64, 8, x)
-#define $u32(x) $(Type_U32, CT_U32, 4, x)
-#define $u64(x) $(Type_U64, CT_U64, 8, x)
-#define $f32(x) $(Type_F32, CT_F32, 4, x)
-#define $f64(x) $(Type_F64, CT_F64, 8, x)
+#define $i32(x) $(Type_I32, CT_I32, 4, (int32_t)x)
+#define $i64(x) $(Type_I64, CT_I64, 8, (int64_t)x)
+#define $u32(x) $(Type_U32, CT_U32, 4, (uint32_t)x)
+#define $u64(x) $(Type_U64, CT_U64, 8, (uint64_t)x)
+#define $f32(x) $(Type_F32, CT_F32, 4, (float)x)
+#define $f64(x) $(Type_F64, CT_F64, 8, (double)x)
 
 #define ct_new_i32() ct_new(&Type_I32)
 #define ct_new_i64() ct_new(&Type_I64)
